feat: Change INTEGER for BIGINT serial id in users

+++ api/user_handler.go
@@ -65,7 +65,7 @@ func (h *UserHandler) HandleGetUser(w http.ResponseWriter, r *http.Request) erro
 		return APIError{Status: http.StatusBadRequest, Msg: "invalid ID", Err: err}
 	}
 
-	user, err := h.userStore.GetUserByID(r.Context(), int32(id))
+	user, err := h.userStore.GetUserByID(r.Context(), int64(id))
 	if errors.Is(err, sql.ErrNoRows) {
 		return APIError{Status: http.StatusBadRequest, Msg: fmt.Sprintf("invalid ID: there are no users with ID %d", id), Err: err}
 	} else if err != nil {
